+++ pywincffi/core/cdefs/headers/structs.h
@@ -73,3 +73,11 @@ typedef struct _WSANETWORKEVENTS {
   long lNetworkEvents;
   int  iErrorCode[...];
 } WSANETWORKEVENTS, *LPWSANETWORKEVENTS;
+
+// https://msdn.microsoft.com/en-us/library/ms684873
+typedef struct _PROCESS_INFORMATION {
+  HANDLE hProcess;
+  HANDLE hThread;
+  DWORD  dwProcessId;
+  DWORD  dwThreadId;
+} PROCESS_INFORMATION, *LPPROCESS_INFORMATION;