+++ docs/source/changelog.rst
@@ -12,6 +12,9 @@ latest
 ~~~~~~
 
 Notable enhancements and changes are:
+
+    * Addition of the :func:`pywincffi.kernel32.process.CreateProcess` and
+      several structures.  Implemented for :issue:`69`.
     * Reworked the test setup steps so they're more consistent.
     * Added a cleanup step to the tests to track down cases that were not
       resetting or testing the Windows API error code.